refactor(domain): separate user and token domains with sub-packages
