+++ src/main/java/woowacourse/auth/application/CustomerService.java
@@ -3,8 +3,8 @@ package woowacourse.auth.application;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import woowacourse.auth.dao.CustomerDao;
-import woowacourse.auth.domain.Customer;
-import woowacourse.auth.domain.Password;
+import woowacourse.auth.domain.user.Customer;
+import woowacourse.auth.domain.user.Password;
 import woowacourse.auth.dto.request.SignUpRequest;
 import woowacourse.auth.dto.request.UpdateMeRequest;
 import woowacourse.auth.dto.request.UpdatePasswordRequest;
