+++ src/test/java/woowacourse/auth/application/AuthServiceTest.java
@@ -11,10 +11,10 @@ import org.junit.jupiter.api.TestInstance.Lifecycle;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.jdbc.Sql;
-import woowacourse.auth.domain.Customer;
-import woowacourse.auth.domain.EncryptedPassword;
-import woowacourse.auth.domain.Password;
-import woowacourse.auth.domain.Token;
+import woowacourse.auth.domain.user.Customer;
+import woowacourse.auth.domain.user.EncryptedPassword;
+import woowacourse.auth.domain.user.Password;
+import woowacourse.auth.domain.token.Token;
 import woowacourse.auth.dto.request.TokenRequest;
 import woowacourse.auth.dto.response.TokenResponse;
 import woowacourse.common.exception.AuthenticationException;
