+++ src/test/java/woowacourse/auth/application/CustomerServiceTest.java
@@ -10,9 +10,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.dao.DataAccessException;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import woowacourse.auth.dao.CustomerDao;
-import woowacourse.auth.domain.Customer;
-import woowacourse.auth.domain.EncryptedPassword;
-import woowacourse.auth.domain.Password;
+import woowacourse.auth.domain.user.Customer;
+import woowacourse.auth.domain.user.EncryptedPassword;
+import woowacourse.auth.domain.user.Password;
 import woowacourse.auth.dto.request.SignUpRequest;
 import woowacourse.auth.dto.request.UpdateMeRequest;
 import woowacourse.auth.dto.request.UpdatePasswordRequest;
