+++ src/test/java/woowacourse/auth/domain/user/PasswordTest.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.domain;
+package woowacourse.auth.domain.user;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatNoException;
@@ -7,6 +7,8 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import woowacourse.auth.domain.user.EncryptedPassword;
+import woowacourse.auth.domain.user.Password;
 import woowacourse.common.exception.InvalidRequestException;
 
 @SuppressWarnings("NonAsciiCharacters")
