+++ src/test/java/woowacourse/util/DatabaseFixture.java
@@ -4,7 +4,7 @@ import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.stereotype.Repository;
-import woowacourse.auth.domain.Customer;
+import woowacourse.auth.domain.user.Customer;
 
 @Repository
 public class DatabaseFixture {
